workers: provide initialized requests.Session to all workers by default
All workers are already doing, and will continue to do some HTTP-based
interactions. Might as well simplify and create the web session in a
central place and reduce custom code in all worker implementations.

As of now, our needs are simple enough so we don't expect workers
wanting to reconfigure their sessions.

Session is not thread-safe depending on the current cookie jar
implementation, so sharing a single object across all threads is out of
the question.